fix: update builder instance management logic
